+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING CHANGE:** `passwordlessLogin` returns object with `verifyUrl` & `expiresAt`
+
 ## [0.0.4] - 2021-04-09
 
 ### Fixed